# src/app/common/utils/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API for dynamic form",
    "description": "Dynamic form api",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3001",
  "basePath": "/api",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    },
    {
      "name": "Session",
      "description": "API for users sign-in the system"
    },
    {
      "name": "Columns",
      "description": "API for create extra column"
    },
    {
      "name": "Forms",
      "description": "API for create a form with a form name And (extra column)"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/users/registration": {
      "post": {
        "tags": ["Users"],
        "summary": "Create new user and send code to user",
        "description": "*Create user with phn or email \n *Response with verification code.Use this code to verify",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/UserModel"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/UserModel"
            }
          }
        }
      }
    },
    "/users/otp_verify": {
      "put": {
        "tags": ["Users"],
        "summary": "Verify otp",
        "description": "New description in bangla\n new another line",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/OtpModel"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Otp updated",
            "schema": {
              "$ref": "#/definitions/OtpModel"
            }
          }
        }
      }
    },
    "/users/password_set": {
      "put": {
        "tags": ["Users"],
        "summary": "Set password after verification",
        "description": "Set password after verification",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/PasswordSetModel"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/SessionModel"
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Create new user in system",
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/SessionModel"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/SessionModel"
            }
          }
        }
      }
    },
    "/users/resend_otp": {
      "put": {
        "tags": ["Users"],
        "summary": "Create new user in system",
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/ResendModel"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/ResendModel"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "UserModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "mdtajalislam1189@gmail.com"
        },
        "phnNo": {
          "type": "string",
          "example": "01883950739"
        },
        "ownGender": {
          "type": "string",
          "example": "Male",
          "required": true
        },
        "findGender": {
          "type": "string",
          "example": "Male",
          "required": true
        }
      }
    },
    "OtpModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "mdtajalislam1189@gmail.com"
        },
        "phnNo": {
          "type": "string",
          "example": "01883950739"
        },
        "verificationCode": {
          "type": "string",
          "example": "123456"
        }
      }
    },
    "SessionModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "mdtajalislam1189@gmail.com"
        },
        "phnNo": {
          "type": "string",
          "example": "01883950739"
        },
        "password": {
          "type": "string",
          "required": true,
          "example": "123456"
        },
        "isRemember": {
          "type": "boolean",
          "example": "false"
        }
      }
    },
    "PasswordSetModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "mdtajalislam1189@gmail.com"
        },
        "phnNo": {
          "type": "string",
          "example": "01883950739"
        },
        "password": {
          "type": "string",
          "example": "123456",
          "required": true
        },
        "passwordConfirmation": {
          "type": "string",
          "examples": "123456"
        },
        "isRemember": {
          "type": "boolean",
          "example": "false"
        }
      }
    },
    "ResendModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "mdtajalislam1189@gmail.com"
        },
        "phnNo": {
          "type": "string",
          "example": "01883950739"
        }
      }
    }
  }
}